requirements.txt
flask
line-bot-sdk
google-cloud-vision
pysqlite3
python-dotenv
pandas